Got it to compile with stack LTS-15.5
Most of the changes are due to the Pretty class that was added, it
seems, in v2.2.0.1-tgt. There were some slightly trickier cases around
some data types that don't seem to be used anymore and changes to the
Package/ GenericPackge records.

Open questions:

- This doesn't pass all the tests. The main failures seem to be around
quoting in certain rendering situations and in error reporting
- Relatedly, how should the new-ish ParseResult monad be incorporated? I
opted for parseGenericPackageDescriptionMaybe because I didn't want to
infect the rest of the module with an API change, but it's unclear how
to exactly recover the old behavior.
- Is the style where it needs to be (whitespace/ indentation, ...)?